Tuning

+++ source/aogmaneo/Encoder.cpp
@@ -48,7 +48,7 @@ void Encoder::forward(
             Int2 iterLowerBound(max(0, fieldLowerBound.x), max(0, fieldLowerBound.y));
             Int2 iterUpperBound(min(vld.size.x - 1, visibleCenter.x + vld.radius), min(vld.size.y - 1, visibleCenter.y + vld.radius));
 
-            int subSum = 0;
+            float subSum = 0.0f;
             int subCount = (iterUpperBound.x - iterLowerBound.x + 1) * (iterUpperBound.y - iterLowerBound.y + 1);
 
             for (int ix = iterLowerBound.x; ix <= iterUpperBound.x; ix++)
@@ -62,7 +62,8 @@ void Encoder::forward(
                     subSum += vl.weights[inCI + vld.size.z * (offset.y + diam * (offset.x + diam * hiddenCellIndex))];
                 }
 
-            sum += static_cast<float>(subSum) / static_cast<float>(max(1, subCount)) * vl.importance;
+            sum += subSum / max(1, subCount) * vl.importance;
+
             totalImportance += vl.importance;
         }
